requirements.txt
os
random
matplotlib.pyplot
import
opencv-python
ultralytics
PIL
oboflow